requirements.txt
# ---- Web Framework ----
Flask==3.0.2
Flask-Session==0.5.0
Werkzeug==3.0.1

# ---- Database ----
mysql-connector-python==8.3.0

# ---- Deep Learning / Vision ----
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
timm==0.9.12

# ---- Image Processing ----
Pillow==10.2.0
opencv-python==4.9.0.80
albumentations==1.4.3
numpy==1.26.4
matplotlib==3.8.2

# ---- PDF Generation ----
reportlab==4.1.0

# ---- Utilities ----
python-dotenv==1.0.1
gunicorn==21.2.0
